Added new variant of <S_+ S_- + S_- S_+>/2 measurement
The new variant reimagines segments by connecting a pair of S+,S- operators while breaking their old connections.
The weight for the reimagined segment is calculated by scaling the old weight by the ratios of the changed J_perps,
and the additional measurements are performed by functional derivatives on the new set of J_perps.